refactor(background): make entrypoint bit more understandable
Less "callback-based programming" style
